LPS-200603 Move RoleModelListener and GroupModelListener into module, and make them as OSGI components instead of spring bean
