+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/model/listener/GroupModelListener.java
@@ -3,33 +3,38 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.announcements.model;
+package com.liferay.announcements.web.internal.model.listener;
 
-import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalServiceUtil;
+import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalService;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.service.ClassNameLocalServiceUtil;
+import com.liferay.portal.kernel.model.ModelListener;
+import com.liferay.portal.kernel.service.ClassNameLocalService;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Christopher Kian
  */
+@Component(service = ModelListener.class)
 public class GroupModelListener extends BaseModelListener<Group> {
 
 	@Override
 	public void onBeforeRemove(Group group) throws ModelListenerException {
 		try {
 			if (group.isSite()) {
-				AnnouncementsEntryLocalServiceUtil.deleteEntries(
+				_announcementsEntryLocalService.deleteEntries(
 					group.getClassNameId(), group.getGroupId());
 			}
 			else {
-				AnnouncementsEntryLocalServiceUtil.deleteEntries(
+				_announcementsEntryLocalService.deleteEntries(
 					group.getClassNameId(), group.getClassPK());
 
 				if (group.isOrganization()) {
-					AnnouncementsEntryLocalServiceUtil.deleteEntries(
-						ClassNameLocalServiceUtil.getClassNameId(Group.class),
+					_announcementsEntryLocalService.deleteEntries(
+						_classNameLocalService.getClassNameId(Group.class),
 						group.getGroupId());
 				}
 			}
@@ -38,5 +43,11 @@ public class GroupModelListener extends BaseModelListener<Group> {
 			throw new ModelListenerException(exception);
 		}
 	}
+
+	@Reference
+	private AnnouncementsEntryLocalService _announcementsEntryLocalService;
+
+	@Reference
+	private ClassNameLocalService _classNameLocalService;
 
 }
