+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/model/listener/RoleModelListener.java
@@ -3,27 +3,35 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.announcements.model;
+package com.liferay.announcements.web.internal.model.listener;
 
-import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalServiceUtil;
+import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalService;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.model.Role;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Christopher Kian
  */
+@Component(service = ModelListener.class)
 public class RoleModelListener extends BaseModelListener<Role> {
 
 	@Override
 	public void onBeforeRemove(Role role) throws ModelListenerException {
 		try {
-			AnnouncementsEntryLocalServiceUtil.deleteEntries(
+			_announcementsEntryLocalService.deleteEntries(
 				role.getClassNameId(), role.getRoleId());
 		}
 		catch (Exception exception) {
 			throw new ModelListenerException(exception);
 		}
 	}
+
+	@Reference
+	private AnnouncementsEntryLocalService _announcementsEntryLocalService;
 
 }
