+++ portal-impl/src/META-INF/model-listener-spring.xml
@@ -18,6 +18,4 @@
 	<bean class="com.liferay.portal.security.permission.UserModelListener" id="com.liferay.portal.security.permission.UserModelListener" />
 	<bean class="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutModelListener" id="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutModelListener" />
 	<bean class="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutSetModelListener" id="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutSetModelListener" />
-	<bean class="com.liferay.portlet.announcements.model.GroupModelListener" id="com.liferay.portlet.announcements.model.GroupModelListener" />
-	<bean class="com.liferay.portlet.announcements.model.RoleModelListener" id="com.liferay.portlet.announcements.model.RoleModelListener" />
 </beans>
